more Wrapped code cleanup
