+++ packages/lesswrong/components/ea-forum/wrapped/EAForumWrappedPage.tsx
@@ -1619,16 +1619,6 @@ const EAForumWrappedPage = ({classes}: {classes: ClassesType}) => {
     year
   })
 
-  // After the user has seen the summary section,
-  // we add a link in the first section to skip down to it for convenience
-  useEffect(() => {
-    if (year === 2023 && entry?.isIntersecting && !currentUser?.wrapped2023Viewed) {
-      void updateCurrentUser({
-        wrapped2023Viewed: true,
-      });
-    }
-  }, [year, entry, updateCurrentUser, currentUser?.wrapped2023Viewed]);
-
   const skipToSummary = useCallback(() => {
     node?.scrollIntoView({ behavior: "smooth", block: "center", inline: "nearest" });
   }, [node]);
@@ -1691,18 +1681,13 @@ const EAForumWrappedPage = ({classes}: {classes: ClassesType}) => {
         <AnalyticsContext pageSectionContext="top">
           <section className={classes.section}>
             <h1 className={classes.heading1}>Your {year} Wrapped</h1>
-            {(year !== 2023 || currentUser.wrapped2023Viewed) &&
-              <button className={classNames(classes.summaryLinkWrapper, classes.skipToSummaryBtn)} onClick={skipToSummary}>
-                Skip to summary
-                <ForumIcon icon="NarrowArrowDown" />
-              </button>
-            }
+            <button className={classNames(classes.summaryLinkWrapper, classes.skipToSummaryBtn)} onClick={skipToSummary}>
+              Skip to summary
+              <ForumIcon icon="NarrowArrowDown" />
+            </button>
             <CloudinaryImage2
               publicId="2023_wrapped"
-              wrapperClassName={classNames(classes.imgWrapper, {
-                [classes.mt60]: currentUser.wrapped2023Viewed,
-                [classes.mt100]: !currentUser.wrapped2023Viewed,
-              })}
+              wrapperClassName={classNames(classes.imgWrapper, classes.mt60)}
               className={classes.img}
             />
           </section>
