+++ packages/lesswrong/lib/collections/users/fragments.ts
@@ -218,8 +218,6 @@ registerFragment(`
     showRecommendedPartners
     hideActiveDialogueUsers
 
-    wrapped2023Viewed
-
     hideSunshineSidebar
   }
 `);
